fix value serde in `patch`, comparison of complex values (#22)
patch and replace were not using jsonToConvex, so if you did db.patch(id, { foo: BigInt(1) }) and then db.get(id) it would throw an error because of the field name "$integer".

also fixes a bug i recently caused where db.patch(id, { foo: null }) would throw an error; whoops.

fix comparisons for complex values like BigInt and arrays, which weren't working, by copying the compare.ts function from the aggregate component. that library isn't perfect, for example i think it considers -0 and +0 to be equal, but it's better than the existing compareValues function which thinks ["a"] < ["a"] and {a:null} < {a: null}

added a test that covers all of the operations for all kinds of values.